+++ G07P3/src/src/utils/Utils.java
@@ -19,7 +19,7 @@ public class Utils {
 	 * 
 	 * @return int
 	 */
-	static int findIntervalAux(ArrayList<Integer> array, int low, int high, Double value)
+	static int findIntervalAux(ArrayList<Double> array, int low, int high, Double value)
 	{
 		if (low == high - 1)
 			return low;
@@ -27,7 +27,7 @@ public class Utils {
 		// Find the middle index
 		int mid = (high + low) / 2;
 
-		if((double)(array.get(mid)) == value)
+		if(array.get(mid) == value)
 			return mid;
 		else if(value < array.get(mid))
 			return findIntervalAux(array, low, mid, value);
@@ -45,7 +45,7 @@ public class Utils {
 	 * 
 	 * @return int
 	 */
-	public static int findInterval(Double value, ArrayList<Integer> array) {
+	public static int findInterval(Double value, ArrayList<Double> array) {
 		int low = 0, high = array.size();
 
 		return findIntervalAux(array, low, high, value);
@@ -54,7 +54,7 @@ public class Utils {
 
 	public static ArrayList<Cromosoma> sortSample(ArrayList<Cromosoma> array){
 		// order from largest to smallest
-    Collections.sort(array, (a, b) -> Integer.compare(b.getFitness(), a.getFitness()));
+    Collections.sort(array, (a, b) -> Double.compare(b.getFitness(), a.getFitness()));
 		return array;
 	}
 }
